WIP: include title_full in default title search to address an issue with square brackets
Because of the way that Aspen normalizes titles that have square brackets
in the 245 field, we can get differences in indexing:

245 $a Title $b [book discussion kit]

gets normalized as

"title book discussion kit"

while

245 $a Title : $b a novel [book discussion kit]

becomes

"title a novel"

Consequently, title searchs for "book discussion kit" would not return
everything that's relevant.

Since title_full includes the full thing, this patch works around the problem
by including the full title in the title search.

More work is needed to sort out all of the desiderata for handling title
content that is enclosed in square brackets. Sometimes you do want
cataloger-supplied text to be indexed, but sometimes (such as GMDs), you
don't, necessarily.

Signed-off-by: Galen Charlton <gmc@equinoxOLI.org>